Refactor: map one Z2M property to multiple capabilities
Add the ability to map one Z2M property to multiple capabilities.

Thanks to this, all color-related code has been cleaned up and is now
centralized in capabilitymap.

This streamlines how we add device support in this project, allowing
future complex devices to be integrated without polluting the codebase.
The only code file that needs to be modified to add a device is now
capabilitymap.